Renamed namespace to the proper name. Added textures for buttons.
UIButton now derives from UIImage which makes more sense.
UIImage and UIText are also considered interactable.
Interactability can be toggled for each object that derives from UIInteractable. (it's off by default except for the UIButton)
Implemented interactions for UIInteractable.
Added buttons and text to the main menu.
Added support for external scripts.
Added script for main menu that loads the main menu scene (entry point for the game).
External scripts can now be attached to GameObjects.
External scripts can now be attached to the MainGame.
Re-order update calls from child objects of GameObject in order for external scripts to have the most updated data for the current frame.
During development scenes will hot-reload when changes are done to the json file allowing for rapid testing of changes to the scene.